\name{idbg.source}
\Rdversion{1.1}
\alias{idbg.source}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{idbg.source}
\description{Source R files and restore breakpoints}
\usage{
idbg.source(fname)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{fname}{R file name to source}
}
\value{ None (invisible NULL).}
\author{ Ronen Kimchi}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{source}}
}
\examples{


\dontrun{idbg.source("my_debugged_R_file.R")}


}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
% \keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
